Output name datatype update
Updated the datatype specification for output image names inside the function file.

+++ scripts/align_z_focus.py
@@ -61,7 +61,7 @@ def get_sms_im_name(time_idx=None,
     if pos_idx is not None:
         im_name += "_p" + str(pos_idx).zfill(int2str_len)
     if slice_idx is not None:
-        im_name += "_z" + str(slice_idx).zfill(int2str_len)
+        im_name += "_z" + str(slice_idx.astype('int64')).zfill(int2str_len)
     if extra_field is not None:
         im_name += "_" + extra_field
     im_name += ext
@@ -159,7 +159,7 @@ def main(input_dir,
         im_name_dst = get_sms_im_name(
             time_idx=meta_row['time_idx'],
             channel_name=meta_row['channel_name'],
-            slice_idx=meta_row['slice_idx_new'].astype('int64'),
+            slice_idx=meta_row['slice_idx_new'],
             pos_idx=meta_row['pos_idx'],
             ext='.tif',
         )
